+++ jdaviz/configs/default/plugins/model_fitting/model_fitting.py
@@ -8,7 +8,7 @@ from specutils.utils import QuantityModel
 from traitlets import Bool, List, Unicode, observe
 from glue.core.data import Data
 
-from jdaviz.core.events import SnackbarMessage
+from jdaviz.core.events import SnackbarMessage, GlobalDisplayUnitChanged
 from jdaviz.core.registries import tray_registry
 from jdaviz.core.template_mixin import (PluginTemplateMixin,
                                         SelectPluginComponent,
@@ -60,6 +60,7 @@ class ModelFitting(PluginTemplateMixin, DatasetSelectMixin,
     * :meth:`create_model_component`
     * :meth:`remove_model_component`
     * :meth:`model_components`
+    * :meth:`valid_model_components`
     * :meth:`get_model_component`
     * :meth:`set_model_component`
     * :meth:`reestimate_model_parameters`
@@ -169,13 +170,17 @@ class ModelFitting(PluginTemplateMixin, DatasetSelectMixin,
         # set the filter on the viewer options
         self._update_viewer_filters()
 
+        self.hub.subscribe(self, GlobalDisplayUnitChanged,
+                           handler=self._on_global_display_unit_changed)
+
     @property
     def user_api(self):
         expose = ['dataset']
         if self.config == "cubeviz":
             expose += ['spatial_subset']
         expose += ['spectral_subset', 'model_component', 'poly_order', 'model_component_label',
-                   'model_components', 'create_model_component', 'remove_model_component',
+                   'model_components', 'valid_model_components',
+                   'create_model_component', 'remove_model_component',
                    'get_model_component', 'set_model_component', 'reestimate_model_parameters',
                    'equation', 'equation_components',
                    'add_results', 'residuals_calculate', 'residuals']
@@ -336,6 +341,7 @@ class ModelFitting(PluginTemplateMixin, DatasetSelectMixin,
         # (won't affect calculations because these locations are masked)
         selected_spec.flux[np.isnan(selected_spec.flux)] = 0.0
 
+        # TODO: can we simplify this logic?
         self._units["x"] = str(
             selected_spec.spectral_axis.unit)
         self._units["y"] = str(
@@ -503,7 +509,37 @@ class ModelFitting(PluginTemplateMixin, DatasetSelectMixin,
         self._initialized_models[comp_label] = initialized_model
 
         new_model["Initialized"] = True
+        new_model["initialized_display_units"] = self._units.copy()
+
+        new_model["compat_display_units"] = True  # always compatible at time of creation
         return new_model
+
+    def _check_model_component_compat(self, axes=['x', 'y'], display_units=None):
+        if display_units is None:
+            display_units = [u.Unit(self._units[ax]) for ax in axes]
+
+        disp_physical_types = [unit.physical_type for unit in display_units]
+
+        for model_index, comp_model in enumerate(self.component_models):
+            compat = True
+            for ax, ax_physical_type in zip(axes, disp_physical_types):
+                comp_unit = u.Unit(comp_model["initialized_display_units"][ax])
+                compat = comp_unit.physical_type == ax_physical_type
+                if not compat:
+                    break
+            self.component_models[model_index]["compat_display_units"] = compat
+
+        # length hasn't changed, so we need to force the traitlet to update
+        self.send_state("component_models")
+        self._check_model_equation_invalid()
+
+    def _on_global_display_unit_changed(self, msg):
+        axis = {'spectral': 'x', 'flux': 'y'}.get(msg.axis)
+
+        # update internal tracking of current units
+        self._units[axis] = str(msg.unit)
+
+        self._check_model_component_compat([axis], [msg.unit])
 
     def remove_model_component(self, model_component_label):
         """
@@ -634,6 +670,9 @@ class ModelFitting(PluginTemplateMixin, DatasetSelectMixin,
         # length hasn't changed, so we need to force the traitlet to update
         self.send_state("component_models")
 
+        # model units may have changed, need to re-check their compatibility with display units
+        self._check_model_component_compat()
+
         # return user-friendly info on revised model
         return self.get_model_component(model_component_label)
 
@@ -645,11 +684,18 @@ class ModelFitting(PluginTemplateMixin, DatasetSelectMixin,
         return [x["id"] for x in self.component_models]
 
     @property
+    def valid_model_components(self):
+        """
+        List of the labels of existing valid (due to display units) model components
+        """
+        return [x["id"] for x in self.component_models if x["compat_display_units"]]
+
+    @property
     def equation_components(self):
         """
         List of the labels of model components in the current equation
         """
-        return re.split('[+*/-]', self.equation.value)
+        return re.split(r'[+*/-]', self.equation.value.replace(' ', ''))
 
     def vue_add_model(self, event):
         self.create_model_component()
@@ -658,10 +704,41 @@ class ModelFitting(PluginTemplateMixin, DatasetSelectMixin,
         self.remove_model_component(event)
 
     @observe('model_equation')
-    def _model_equation_changed(self, event):
+    def _check_model_equation_invalid(self, event=None):
         # Length is a dummy check to test the infrastructure
         if len(self.model_equation) == 0:
-            self.model_equation_invalid_msg = 'model equation is required'
+            self.model_equation_invalid_msg = 'model equation is required.'
+            return
+        if '' in self.equation_components:
+            # includes an operator without a variable (ex: 'C+')
+            self.model_equation_invalid_msg = 'incomplete equation.'
+            return
+
+        components_not_existing = [comp for comp in self.equation_components
+                                   if comp not in self.model_components]
+        if len(components_not_existing):
+            if len(components_not_existing) == 1:
+                msg = "is not an existing model component."
+            else:
+                msg = "are not existing model components."
+            self.model_equation_invalid_msg = f'{", ".join(components_not_existing)} {msg}'
+            return
+        components_not_valid = [comp for comp in self.equation_components
+                                if comp not in self.valid_model_components]
+        if len(components_not_valid):
+            if len(components_not_valid) == 1:
+                msg = ("is currently disabled because it has"
+                       " incompatible units with the current display units."
+                       " Remove the component from the equation,"
+                       " re-estimate its free parameters to use the new units"
+                       " or revert the display units.")
+            else:
+                msg = ("are currently disabled because they have"
+                       " incompatible units with the current display units."
+                       " Remove the components from the equation,"
+                       " re-estimate their free parameters to use the new units"
+                       " or revert the display units.")
+            self.model_equation_invalid_msg = f'{", ".join(components_not_valid)} {msg}'
             return
         self.model_equation_invalid_msg = ''
 
@@ -707,6 +784,8 @@ class ModelFitting(PluginTemplateMixin, DatasetSelectMixin,
         if not self.spectral_subset_valid:
             valid, spec_range, subset_range = self._check_dataset_spectral_subset_valid(return_ranges=True)  # noqa
             raise ValueError(f"spectral subset '{self.spectral_subset.selected}' {subset_range} is outside data range of '{self.dataset.selected}' {spec_range}")  # noqa
+        if len(self.model_equation_invalid_msg):
+            raise ValueError(f"model equation is invalid: {self.model_equation_invalid_msg}")
 
         if self.cube_fit:
             ret = self._fit_model_to_cube(add_data=add_data)