+++ jdaviz/configs/default/plugins/subset_plugin/subset_plugin.py
@@ -2,6 +2,7 @@ import os
 
 import numpy as np
 
+import astropy.units as u
 from glue.core.message import EditSubsetMessage, SubsetUpdateMessage
 from glue.core.edit_subset_mode import (AndMode, AndNotMode, OrMode,
                                         ReplaceMode, XorMode)
@@ -12,8 +13,7 @@ from glue_jupyter.widgets.subset_mode_vuetify import SelectionModeMenu
 from glue_jupyter.common.toolbar_vuetify import read_icon
 from traitlets import Any, List, Unicode, Bool, observe
 
-
-from jdaviz.core.events import SnackbarMessage
+from jdaviz.core.events import SnackbarMessage, GlobalDisplayUnitChanged
 from jdaviz.core.registries import tray_registry
 from jdaviz.core.template_mixin import PluginTemplateMixin, DatasetSelectMixin, SubsetSelect
 
@@ -65,12 +65,15 @@ class SubsetPlugin(PluginTemplateMixin, DatasetSelectMixin):
                                    handler=self._sync_selected_from_state)
         self.session.hub.subscribe(self, SubsetUpdateMessage,
                                    handler=self._on_subset_update)
+        self.session.hub.subscribe(self, GlobalDisplayUnitChanged,
+                                   handler=self._on_display_unit_changed)
 
         self.subset_select = SubsetSelect(self,
                                           'subset_items',
                                           'subset_selected',
                                           default_text="Create New")
         self.subset_states = []
+        self.spectral_display_unit = None
 
     def _sync_selected_from_state(self, *args):
         if not hasattr(self, 'subset_select'):
@@ -140,7 +143,8 @@ class SubsetPlugin(PluginTemplateMixin, DatasetSelectMixin):
         -------
 
         """
-        subset_information = self.app.get_subsets(self.subset_selected, simplify_spectral=False)
+        subset_information = self.app.get_subsets(self.subset_selected, simplify_spectral=False,
+                                                  use_display_units=True)
         _around_decimals = 6  # Avoid 30 degrees from coming back as 29.999999999999996
         if not subset_information:
             return
@@ -188,10 +192,12 @@ class SubsetPlugin(PluginTemplateMixin, DatasetSelectMixin):
                 subset_type = subset_state.roi.__class__.__name__
 
             elif isinstance(subset_state, RangeSubsetState):
-                lo = subset_state.lo
-                hi = subset_state.hi
-                subset_definition = [{"name": "Lower bound", "att": "lo", "value": lo, "orig": lo},
-                                     {"name": "Upper bound", "att": "hi", "value": hi, "orig": hi}]
+                lo = spec['region'].lower
+                hi = spec['region'].upper
+                subset_definition = [{"name": "Lower bound", "att": "lo", "value": lo.value,
+                                      "orig": lo.value, "unit": str(lo.unit)},
+                                     {"name": "Upper bound", "att": "hi", "value": hi.value,
+                                      "orig": hi.value, "unit": str(hi.unit)}]
                 subset_type = "Range"
             if len(subset_definition) > 0:
                 # Note: .append() does not work for List traitlet.
@@ -221,6 +227,7 @@ class SubsetPlugin(PluginTemplateMixin, DatasetSelectMixin):
 
         self._unpack_get_subsets_for_ui()
 
+<<<<<<< HEAD
     def vue_simplify_subset(self, *args):
         if len(self.subset_states) < 2:
             self.hub.broadcast(SnackbarMessage("Cannot simplify spectral subset "
@@ -230,6 +237,14 @@ class SubsetPlugin(PluginTemplateMixin, DatasetSelectMixin):
         att = self.subset_states[0].att
         self.app.simplify_spectral_subset(subset_name=self.subset_selected, att=att,
                                           overwrite=True)
+=======
+    def _on_display_unit_changed(self, msg):
+        # We only care about the spectral units, since flux units don't affect spectral subsets
+        if msg.axis == "spectral":
+            self.spectral_display_unit = msg.unit
+            if self.subset_selected != self.subset_select.default_text:
+                self._get_subset_definition(self.subset_selected)
+>>>>>>> a9151a48acb62b44db69c360245bd46c6c8b7a9a
 
     def vue_update_subset(self, *args):
         status, reason = self._check_input()
@@ -246,6 +261,16 @@ class SubsetPlugin(PluginTemplateMixin, DatasetSelectMixin):
                     d_val = np.radians(d_att["value"])
                 else:
                     d_val = float(d_att["value"])
+
+                # Convert from display unit to original unit if necessary
+                if self.subset_types[index] == "Range":
+                    if self.spectral_display_unit is not None:
+                        x_att = sub_states.att
+                        base_units = self.app.data_collection[0].get_component(x_att).units
+                        if self.spectral_display_unit != base_units:
+                            d_val = d_val*u.Unit(self.spectral_display_unit)
+                            d_val = d_val.to(u.Unit(base_units))
+                            d_val = d_val.value
 
                 if float(d_att["orig"]) != d_val:
                     if self.subset_types[index] == "Range":